Web Worker now sends and receives correctly
Just a decode issue now when sending the message back

+++ src/main/scala/roguelike/model/DungeonGenConfig.scala
@@ -3,5 +3,5 @@ package roguelike.model
 import scala.scalajs.js
 
 trait DungeonGenConfig extends js.Object:
-  val seed: Long
+  val seed: Double
   val currentLevel: Int
